+++ glados.cabal
@@ -27,8 +27,8 @@ library
   exposed-modules:
       Ast
       DefaultSymbol
-      Lambda
       Lib
+      Types
   other-modules:
       Paths_glados
   hs-source-dirs:
